feat(reinforce-learning): fully removed environment
